Fix acceptance tests
- Adds submodule`terraform-git-module-acctest` to `internal/git/testdata` dir
- Updates Actions `go-test` job to checkout submodules
- Update tests for `git_repository` data source to use submodule
- Makes some minor adjustments
