+++ .github/workflows/test.yml
@@ -26,7 +26,7 @@ jobs:
     steps:
     - uses: actions/checkout@v2
       with:
-        fetch-depth: 0
+        submodules: true
     - uses: actions/setup-go@v2-beta
       with:
         go-version: '~1.14'
